requirements.txt
torch==1.9.0
transformers.egg==info
dataclasses==0.8
PyYAML==5.4.1
transformers==4.10.3
